Possible overflow fix
If a file is too large the frequency counters could overflow.

+++ Huffman/src/huffman_tree.h
@@ -29,7 +29,7 @@ typedef struct ascii_huffman {
     /**
      * Counts the frequency of every symbol
      */
-    unsigned int charFreq[256];
+    unsigned long int charFreq[256];
 
 } ASCIIHuffman;
